+++ workers/ohsome_quality_analyst/api/response_models.py
@@ -8,6 +8,8 @@ from ohsome_quality_analyst.api.request_models import (
 )
 from ohsome_quality_analyst.definitions import ATTRIBUTION_URL
 from ohsome_quality_analyst.indicators.models import IndicatorMetadata
+from ohsome_quality_analyst.quality_dimensions.definitions import QualityDimensionEnum
+from ohsome_quality_analyst.quality_dimensions.models import QualityDimension
 from ohsome_quality_analyst.reports.models import ReportMetadata
 from ohsome_quality_analyst.topics.models import TopicDefinition
 from ohsome_quality_analyst.utils.helper import snake_to_hyphen
@@ -33,6 +35,18 @@ class TopicMetadataResponse(ResponseBase):
         for key in value.keys():
             TopicEnum(key)
         return value
+
+
+class QualityDimensionMetadataResponse(ResponseBase):
+    result: dict[str, QualityDimension]
+
+    @validator("result")
+    @classmethod
+    def check_quality_dimension_dict(cls, values):
+        assert len(values) > 0
+        for key in values.keys():
+            QualityDimensionEnum(key)
+        return values
 
 
 class IndicatorMetadataResponse(ResponseBase):
@@ -64,5 +78,11 @@ class MetadataResponse(ResponseBase):
         indicators: dict[str, IndicatorMetadata]
         reports: dict[str, ReportMetadata]
         topics: dict[str, TopicDefinition]
+        quality_dimensions: dict[str, QualityDimension]
+
+        class Config:
+            alias_generator = snake_to_hyphen
+            frozen = True
+            extra = "forbid"
 
     result: MetadataResultSchema